Update ST QA
